submission_pool/baseline-RL/requirements.txt
dm-tree==0.1.6
gym==0.23.1
numpy==1.22.3
supersuit==3.3.5
torch==1.10.0
pandas==1.4.1
matplotlib==3.5.1
git+http://gitlab.aicrowd.com/neural-mmo/neurips2022-nmmo.git